Feat: Migrate to latest g2o and opencv 4 for ch7

+++ ch7/orb_self.cpp
@@ -397,10 +397,10 @@ void ComputeORB(const cv::Mat &img, vector<cv::KeyPoint> &keypoints, vector<Desc
 void BfMatch(const vector<DescType> &desc1, const vector<DescType> &desc2, vector<cv::DMatch> &matches) {
   const int d_max = 40;
 
-  for (size_t i1 = 0; i1 < desc1.size(); ++i1) {
+  for (int i1 = 0; i1 < desc1.size(); ++i1) {
     if (desc1[i1].empty()) continue;
     cv::DMatch m{i1, 0, 256};
-    for (size_t i2 = 0; i2 < desc2.size(); ++i2) {
+    for (int i2 = 0; i2 < desc2.size(); ++i2) {
       if (desc2[i2].empty()) continue;
       int distance = 0;
       for (int k = 0; k < 8; k++) {
